[Form] Removed remaining deprecation notices in the test suite
